+++ components/mdz/pkg/cmd/organization/list.go
@@ -19,11 +19,26 @@ type factoryOrganizationList struct {
 	repoOrganization repository.Organization
 	Limit            int
 	Page             int
+	SortOrder        string
+	StartDate        string
+	EndDate          string
 	JSON             bool
 }
 
 func (f *factoryOrganizationList) runE(cmd *cobra.Command, _ []string) error {
-	orgs, err := f.repoOrganization.Get(f.Limit, f.Page)
+	if len(f.StartDate) > 0 {
+		if err := utils.ValidateDate(f.StartDate); err != nil {
+			return err
+		}
+	}
+
+	if len(f.EndDate) > 0 {
+		if err := utils.ValidateDate(f.EndDate); err != nil {
+			return err
+		}
+	}
+
+	orgs, err := f.repoOrganization.Get(f.Limit, f.Page, f.SortOrder, f.StartDate, f.EndDate)
 	if err != nil {
 		return err
 	}
@@ -84,6 +99,12 @@ func (f *factoryOrganizationList) setFlags(cmd *cobra.Command) {
 		"Specifies the number of organizations to retrieve per page")
 	cmd.Flags().IntVar(&f.Page, "page", 1,
 		"Specifies the page number for paginated results")
+	cmd.Flags().StringVar(&f.SortOrder, "sort-order", "",
+		"Specifies the sort order for results (e.g., 'asc' for ascending, 'desc' for descending)")
+	cmd.Flags().StringVar(&f.StartDate, "start-date", "",
+		"Specifies the start date for filtering results (format: YYYY-MM-DD)")
+	cmd.Flags().StringVar(&f.EndDate, "end-date", "",
+		"Specifies the end date for filtering results (format: YYYY-MM-DD)")
 	cmd.Flags().BoolP("help", "h", false, "Displays more information about the Mdz CLI")
 }
 
@@ -107,6 +128,7 @@ func newCmdOrganizationList(f *factoryOrganizationList) *cobra.Command {
 			"$ mdz organization list --limit 10",
 			"$ mdz organization list --page 3",
 			"$ mdz organization list --limit 10 --page 2",
+			"$ mdz organization list --sort-order desc --start-date 2024-12-10 --end-date 2024-12-20",
 		),
 		RunE: f.runE,
 	}