+++ components/mdz/pkg/cmd/portfolio/list.go
@@ -22,6 +22,9 @@ type factoryPortfolioList struct {
 	LedgerID       string
 	Limit          int
 	Page           int
+	SortOrder      string
+	StartDate      string
+	EndDate        string
 	JSON           bool
 }
 
@@ -44,7 +47,20 @@ func (f *factoryPortfolioList) runE(cmd *cobra.Command, _ []string) error {
 		f.LedgerID = id
 	}
 
-	portfolios, err := f.repoPortfolio.Get(f.OrganizationID, f.LedgerID, f.Limit, f.Page)
+	if len(f.StartDate) > 0 {
+		if err := utils.ValidateDate(f.StartDate); err != nil {
+			return err
+		}
+	}
+
+	if len(f.EndDate) > 0 {
+		if err := utils.ValidateDate(f.EndDate); err != nil {
+			return err
+		}
+	}
+
+	portfolios, err := f.repoPortfolio.Get(f.OrganizationID, f.LedgerID, f.Limit, f.Page,
+		f.SortOrder, f.StartDate, f.EndDate)
 	if err != nil {
 		return err
 	}
@@ -101,6 +117,12 @@ func (f *factoryPortfolioList) setFlags(cmd *cobra.Command) {
 		"Specifies the number of ledgers to retrieve per page")
 	cmd.Flags().IntVar(&f.Page, "page", 1,
 		"Specifies the page number for paginated results")
+	cmd.Flags().StringVar(&f.SortOrder, "sort-order", "",
+		"Specifies the sort order for results (e.g., 'asc' for ascending, 'desc' for descending)")
+	cmd.Flags().StringVar(&f.StartDate, "start-date", "",
+		"Specifies the start date for filtering results (format: YYYY-MM-DD)")
+	cmd.Flags().StringVar(&f.EndDate, "end-date", "",
+		"Specifies the end date for filtering results (format: YYYY-MM-DD)")
 	cmd.Flags().BoolP("help", "h", false, "Displays more information about the Mdz CLI")
 }