+++ components/mdz/pkg/cmd/portfolio/list_test.go
@@ -2,9 +2,10 @@ package portfolio
 
 import (
 	"bytes"
-	"go.uber.org/mock/gomock"
 	"testing"
 	"time"
+
+	"go.uber.org/mock/gomock"
 
 	"github.com/LerianStudio/midaz/components/mdz/internal/domain/repository"
 	"github.com/LerianStudio/midaz/components/mdz/pkg/factory"
@@ -85,7 +86,8 @@ func Test_newCmdPortfolioList(t *testing.T) {
 	}
 
 	mockRepo.EXPECT().Get(
-		gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
+		gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(),
+		gomock.Any(), gomock.Any(), gomock.Any()).
 		Return(list, nil)
 
 	err := cmd.Execute()